requirements.txt
blinker==1.5
certifi==2022.6.15
click==8.1.3
elastic-transport==8.1.2
elasticsearch==8.3.1
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.2
SQLAlchemy==1.4.39
urllib3==1.26.10
Werkzeug==2.1.2
python-dotenv==0.20.0
